requirements.txt
fastapi
uvicorn
docxtpl
docx2txt
python-docx
Pillow
python-multipart

flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
gunicorn==20.1.0
openai==0.28.0
python-dotenv==1.0.0
python-docx==0.8.11
python-docx==0.8.11
Pillow==10.0.0
=======
fastapi
uvicorn
docxtpl
docx2txt
python-docx
Pillow
python-multipart
>>>>>>> 76daf16 (Initial commit: backend cardiology for cloud deploy)
